[internal] Fix pep561_integration_test.py to be hermetic to new Pants releases (Cherry-pick of #12961)
`pantsbuild.pants.testutil` depends on `pantsbuild.pants`. Our test for testutil wasn't including the wheel for pantsbuild.pants from the filesystem, so it was falling back to PyPI to satisfy that dep. That works when the wheel is released to PyPI, but caused our new release of Pants 2.7.0rc3 to fail as the wheel is not yet published.

[ci skip-rust]
[ci skip-build-wheels]
